chore: update deps and pause development
i have updated all taliyah dependencies to their latest compatible
versions according to the semantic versioning system, to help keep
taliyah free of security vulnerabilities in terms of the third party
dependencies used. i am also announcing a pause in development for the
forseeable future. more details regarding this are below.

no major changes to taliyah are expected for the forseeable future, due
to lack of motivation to maintain this project, and the fact that poise
has not seen any major releases since january. while serenity has seen
some new releases, there has not been a major release of serenity in
quite a while as well. while i am not completely ceasing development of
taliyah and will continue to update dependencies as necessary to keep
the project free of security vulnerabilities, no new feature development
is likely to occur for at least the short term.

it should also be noted that another reason i am pausing development is
due to a significant lack of an active userbase and the fact that
taliyah in its current state offers no worthwhile features over more
popular, more refined, and more maintained options. i started the poise
rewrite due to the discontinuation of serenity's command framework, the
framework that taliyah originally used, but in complete honesty, i have
rewritten taliyah so many times, that it has just not become fun for me
anymore. it has become more of a drag than something i actually
currently like doing. i first wrote taliyah in javascript before later
moving it over to typescript back in early 2019, before starting to
rewrite it again in rust in late 2019. and then pretty much again in
june 2024 (still in rust) when i decided to port taliyah to use the
poise command framework.

however, all of these rewrites have taken a toll on my desire and drive
to further maintain taliyah. this and more are why i have decided to
pause major feature development of taliyah, however like i said i will
continue to keep taliyah's dependencies up to date, to ensure as much
security and stability as possible. however, if and when serenity /
poise get updated, i likely will not be porting taliyah over to their
new versions, taliyah will stay on serenity 0.12.x and poise 0.6.x for
the forseeable future, unless my stance on this matter changes.

i still want to continue to hone and refine my programming skills in
rust, so i will likely pursue other ideas, but they won't be instantly
made public like my previous projects, such as this one, have.

to anyone who has followed taliyah's development, and at all used
taliyah in the past, thank you. i am sorry that i am pausing taliyah's
development, but this was pretty much already the case due to the rapid
slow down of development that has happened, as observed from the fact
that aside fom dependency updates, no major changes to the codebase have
been made since (roughly) august 3rd, and even before that, i paused
taliyah's development for over a year and a half back in november 2022.

i should once again reiterate that this is *not* the end of taliyah's
development for good, and i may eventually come back to the project when
and if my drive and desire to work on taliyah comes back. i will draft
another commit shortly that updates the README to mention the
development pause.